chore: moved folders according to the new structure - Moved folders to features according to the new structure of the IT-INCUBATOR STYLE GUIDE project.
