+++ src/pages/create/index.tsx
@@ -1,5 +1,5 @@
-import { Create } from '@/features/ui/userHub'
 import { getLayout } from '@/common/components'
+import { Create } from '@/features/userHub/ui'
 
 Create.getLayout = getLayout
 export default Create
