+++ src/pages/createNewPassword/index.tsx
@@ -1,5 +1,5 @@
 import { getLayout } from '@/common/components'
-import { CreateNewPassword } from '@/features/ui/auth'
+import { CreateNewPassword } from '@/features/auth/ui'
 
 CreateNewPassword.getLayout = getLayout
 export default CreateNewPassword
