+++ src/pages/favorites/index.tsx
@@ -1,5 +1,5 @@
-import { Favorites } from '@/features/ui/userHub'
 import { getLayout } from '@/common/components'
+import { Favorites } from '@/features/userHub/ui'
 
 Favorites.getLayout = getLayout
 export default Favorites
