+++ src/pages/forgotPassword/index.tsx
@@ -1,5 +1,5 @@
 import { getLayout } from '@/common/components'
-import { ForgotPassword } from '@/features/ui/auth'
+import { ForgotPassword } from '@/features/auth/ui'
 
 ForgotPassword.getLayout = getLayout
 export default ForgotPassword
