+++ src/pages/messenger/index.tsx
@@ -1,5 +1,5 @@
-import { Messenger } from '@/features/ui/userHub'
 import { getLayout } from '@/common/components'
+import { Messenger } from '@/features/userHub/ui'
 
 Messenger.getLayout = getLayout
 export default Messenger
