+++ src/pages/myProfile/index.tsx
@@ -1,5 +1,5 @@
-import { MyProfile } from '@/features/ui/userHub'
 import { getLayout } from '@/common/components'
+import { MyProfile } from '@/features/userHub/ui'
 
 MyProfile.getLayout = getLayout
 export default MyProfile
