+++ src/pages/passwordRecovery/index.tsx
@@ -1,5 +1,5 @@
 import { getLayout } from '@/common/components'
-import { PasswordRecovery } from '@/features/ui/auth'
+import { PasswordRecovery } from '@/features/auth/ui'
 
 PasswordRecovery.getLayout = getLayout
 export default PasswordRecovery
