+++ src/pages/search/index.tsx
@@ -1,5 +1,5 @@
 import { getLayout } from '@/common/components'
-import { Search } from '@/features/ui/userHub'
+import { Search } from '@/features/userHub/ui'
 
 Search.getLayout = getLayout
 export default Search
