+++ src/pages/signIn/index.tsx
@@ -1,5 +1,5 @@
 import { getLayout } from '@/common/components'
-import { SignIn } from '@/features/ui/auth'
+import { SignIn } from '@/features/auth/ui'
 
 SignIn.getLayout = getLayout
 export default SignIn
