+++ src/pages/signUp/index.tsx
@@ -1,5 +1,5 @@
 import { getLayout } from '@/common/components'
-import { SignUp } from '@/features/ui/auth'
+import { SignUp } from '@/features/auth/ui'
 
 SignUp.getLayout = getLayout
 export default SignUp
