+++ src/pages/statistics/index.tsx
@@ -1,5 +1,5 @@
-import { Statistics } from '@/features/ui/userHub'
 import { getLayout } from '@/common/components'
+import { Statistics } from '@/features/userHub/ui'
 
 Statistics.getLayout = getLayout
 export default Statistics
